requirements.txt
python-dotenv
autopep8
discord.py
pytz